Move ServiceContainer#canCreateWrappedService
