+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/handler/ComponentDescriptor.kt
@@ -14,7 +14,7 @@ import io.github.freya022.botcommands.internal.core.BContextImpl
 import io.github.freya022.botcommands.internal.core.options.OptionType
 import io.github.freya022.botcommands.internal.core.options.builder.CustomOptionBuilderImpl
 import io.github.freya022.botcommands.internal.core.reflection.toMemberParamFunction
-import io.github.freya022.botcommands.internal.core.service.provider.canCreateWrappedService
+import io.github.freya022.botcommands.internal.core.service.canCreateWrappedService
 import io.github.freya022.botcommands.internal.options.transformParameters
 import io.github.freya022.botcommands.internal.parameters.OptionParameter
 import io.github.freya022.botcommands.internal.parameters.ResolverContainer
